% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_scale_with_blueprint.R
\name{build_scale_with_blueprint}
\alias{build_scale_with_blueprint}
\title{Setup an initial FC scale that is consistent with the desired blueprint.}
\usage{
TO BE DONE
}
\arguments{
\item{item_df}{Data frame containing information related to all the available items}

\item{blueprint}{Pre-specified blueprint for your blocks. Preferably constructed from \code{construct_blueprint()}}

\item{bp_block_name, bp_item_nums_name, bp_trait_name, bp_sign_name}{Column names in the blueprint that specifies block number, item number in the block, desired trait of the item, and desired keying of the item, respectively}

\item{bp_matching_criterion_name}{Column name in the blueprint that indicates the additional matching criterion (cutoff value) you wish to test}

\item{df_item_nums_name, df_trait_name, df_sign_name, }{Column names in \code{item_df} that specifies item_number, trait of the item, and keying of the item, respectively}

\item{df_matching_criterion_name}{Optional. Column name in \code{item_df} that is used to evaluate the matching criterion specified in \code{bp_matching_criterion_name}}

\item{df_matching_function}{Optional. A character string containing function name for evaluating the matching criterion}

\item{df_matching_adjust_factor}{Optional. A numeric value. If after \code{max_attempts_in_comb} attempts the additional
criteria in \code{df_matching_criterion_name} cannot be met (> the cutoff value specified in \code{blueprint[, bp_matching_criterion_name]}), 
multiply that cutoff value by this adjusting factor.}

\item{max_attempts_in_comb}{An integer value. How many attempts will be made for finding a block that satisfies the blueprint, before we adjust the cutoff value?}

\item{max_attempts_in_adjust}{An integer value. How many attempts will be made for adjusting cutoff value? 
Will throw a warning and return the currently partially constructed scale (and specify which block might have problems) if number of attempts exceeds this value.}
}
\value{
TO BE DONE
}
\description{
TO BE DONE
}
\details{
Although automatically finding the block combinations that can satisfy
multiple certain criteria for matching can be helpful, users may also wish to
have exact specifications for some blocks in many cases. This function serves as this purpose.
Typically in FC construction, we may want to explicitly specify the trait and keying combinations
so this function allows you to explicitly do that. Users are free to extend this function if further
exact specifications are needed.

For now, this function also allows users to specify one additional matching criterion for
the blocks. Users can designate the function for calculating this criterion and specify
a multiplicative adjusting factor, if the criterion fails to be met after a specified number of attempts.
One good example of matching criterion is matching in rating, where you want ratings of the items in the same block
to be less than a certain cutoff.
}
\examples{
TO BE DONE

}
\author{
Mengtong Li
}
