% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_target_dist.R
\name{build_target_dist}
\alias{build_target_dist}
\title{Create Target Pair Distribution for Scale-Level Fit}
\usage{
build_target_dist(
  traits,
  total_pairs,
  equal_mixed_ratio = c(1, 1),
  allow_same_trait = FALSE
)
}
\arguments{
\item{traits}{A character vector of unique trait names (e.g., c("O", "C", "E", "A", "N")).}

\item{total_pairs}{Integer. The total number of pairs your test will generate.}

\item{equal_mixed_ratio}{Numeric vector of length 2. The desired ratio of
"equal" keyed pairs to "mixed" keyed pairs (e.g., c(1, 1) for 50/50 split).}

\item{allow_same_trait}{Logical. Do we allow items measuring the same traits to be paired together?}
}
\value{
A flexible data frame containing target counts for the SA algorithm.
}
\description{
Create Target Pair Distribution for Scale-Level Fit
}
