% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TIRT_var_names.R
\name{build_TIRT_var_names}
\alias{build_TIRT_var_names}
\title{Automatically build variable names for the pairwise/rank responses in the TIRT model}
\usage{
TODO
}
\arguments{
\item{item_name}{The prefix you want to have for your response variables.}

\item{block_size, N_blocks}{The block size and total number of the forced-choice scale.}

\item{format}{What format should the converted responses be in? Can be \code{"pairwise"} or \code{"ranks"}.}
}
\value{
A vector of variable names
}
\description{
TODO
}
\details{
TODO
}
\examples{
build_var_names("i", block_size = 3, N_blocks = 20, format = "pairwise")
build_var_names("i", block_size = 5, N_blocks = 12, format = "ranks")


}
\author{
Mengtong Li
}
