% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_block_energy.R
\name{cal_block_energy}
\alias{cal_block_energy}
\title{Fast Calculation of Item Block "Energy"}
\usage{
cal_block_energy(block, char_list, weights, fun_list)
}
\arguments{
\item{block}{An \emph{n} by \emph{k} integer matrix of item indices.}

\item{char_list}{A list of vectors, where each vector contains the full item pool
data for one specific characteristic.}

\item{weights}{A numeric vector of weights.}

\item{fun_list}{A list of pre-matched R function objects (e.g., \code{list(facfun, var)}).}
}
\value{
A numeric value indicating the total energy for the given item blocks.
}
\description{
Calculates the total "energy" of one or multiple paired item blocks.
This function has been heavily optimized for speed, as it is called thousands
of times during the Simulated Annealing optimization loop.
}
