% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_block_energy_with_iia.R
\name{cal_block_energy_with_iia}
\alias{cal_block_energy_with_iia}
\title{Fast Calculation of Item Block "Energy" with Inter-Item Agreement}
\usage{
cal_block_energy_with_iia(
  block,
  char_list,
  weights,
  fun_list,
  rater_chars,
  iia_weights = c(BPlin = 1, BPquad = 1, AClin = 1, ACquad = 1),
  verbose = FALSE
)
}
\arguments{
\item{block}{An \emph{n} by \emph{k} integer matrix of item indices.}

\item{char_list}{A list of vectors for item characteristics.}

\item{weights}{A numeric vector of weights for item characteristics.}

\item{fun_list}{A list of pre-matched R function objects.}

\item{rater_chars}{A matrix of participant responses.}

\item{iia_weights}{A numeric vector of length 4 for IIA metrics.}

\item{verbose}{Logical. If TRUE, prints metrics. Defaults to FALSE.}
}
\value{
A numeric value indicating the total energy.
}
\description{
Calculates local block energy including IIA metrics. Heavily
optimized for Simulated Annealing.
}
