% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_TIRT_response.R
\name{convert_to_TIRT_response}
\alias{convert_to_TIRT_response}
\title{Convert the latent utility values into Thurstonian IRT pairwise/rank responses
with a pre-specified block design}
\usage{
TO BE DONE
}
\arguments{
\item{Utility}{The utility matrix of all items.}

\item{block_design}{A numeric matrix specifying which items will be in the same forced-choice block (row).}

\item{format}{What format should the converted responses be in? Can be \code{"pairwise"} or \code{"ranks"}.}

\item{partial}{Only used when \code{format == "ranks"}. Should partial ranking responses be produced?}

\item{block_size, N_blocks}{The block size and total number of the forced-choice scale. 
Preferably left blank and obtained through \code{block_design}.}

\item{N_response}{Number of simulated responses you wish to generate. Default to \code{nrow(Utility)}.}
}
\value{
A data frame containing pairwise (if \code{format == "pairwise"}) or
rank (if \code{format == "ranks"}) responses to each block for the \code{N_response} participants.
}
\description{
TO BE DONE
}
\details{
TODO
}
\examples{
TO BE DONE


}
\author{
Mengtong Li
}
