% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tirt_stan_scores.R
\name{extract_tirt_stan_scores}
\alias{extract_tirt_stan_scores}
\title{Extract Trait Scores and SEs from a Fitted Stan TIRT Model}
\usage{
extract_tirt_stan_scores(fit, stan_data)
}
\arguments{
\item{fit}{A fitted cmdstanr object.}

\item{stan_data}{The data list passed to Stan (must contain the 'trait_names' attribute).}
}
\value{
A data frame with respondents as rows, and Traits and Trait SEs as columns.
}
\description{
Extract Trait Scores and SEs from a Fitted Stan TIRT Model
}
