% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_item_info.R
\docType{data}
\name{FC_item_info}
\alias{FC_item_info}
\title{Keying and factor information for HEXACO-60 items}
\format{
A data frame with 60 rows and 3 columns
\describe{
\item{Item ID}{Item number}
\item{keying}{Keying of each item}
\item{factor}{Measured factor of each item}
}
}
\usage{
FC_item_info
}
\description{
Includes the keying and measured factor of the 60 HEXACO-60 items.
}
\references{
Ashton, M. C., & Lee, K. (2009). The HEXACO-60: A short measure of the major dimensions of personality.
\emph{Journal of Personality Assessment, 91}(4), 340-345. \doi{10.1080/00223890902935878}

Li, M., Zhang, B., Li, L., Sun, T., & Brown, A. (2025). Mixed-keying or desirability-matching in the
construction of forced-choice measures? An empirical investigation and practical recommendations.
\emph{Organizational Research Methods, 28}(2), 296-329. \doi{10.1177/10944281241229784}
}
\keyword{datasets}
