% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tirt_lavaan_syntax.R
\name{generate_tirt_lavaan_syntax}
\alias{generate_tirt_lavaan_syntax}
\title{Generate lavaan Syntax for Thurstonian Models}
\usage{
generate_tirt_lavaan_syntax(
  n_blocks,
  block_size,
  key_matrix,
  trait_col,
  key_col,
  cor_matrix = NULL,
  model_type = c("TFM", "TIRT"),
  force_positive_variances = TRUE,
  more_constraints = ""
)
}
\arguments{
\item{n_blocks}{Integer. Number of blocks in the questionnaire.}

\item{block_size}{Integer. Number of items per block.}

\item{key_matrix}{A data.frame with two columns, one indicating item trait,
another indicating item sign.}

\item{trait_col}{Character string. The name of the column in \code{key_matrix} indicating the trait
measured by the item.}

\item{key_col}{Character string. The name of the column in \code{key_matrix} indicating the keying
direction of the item (e.g., positive/negative).}

\item{cor_matrix}{Optional matrix. Starting values for trait correlations.}

\item{model_type}{Character. Either "TFM" (Second-Order Thurstonian Factor Model) or "IRT" (First-Order).}

\item{force_positive_variances}{Logical. If TRUE, prevents Heywood cases.
In "Factor" models, utility variances > 0.0001.
In "IRT" models, utility variances and derived correlated errors > 0.0001.
Defaults to TRUE.}

\item{more_constraints}{Additional lavaan syntax strings for constraining certain parameters if needed.}
}
\value{
A character string containing the lavaan model syntax.
}
\description{
Generates the syntax required to fit either a Second-Order Thurstonian Factor
Model or a First-Order Thurstonian IRT Model in lavaan.
}
