% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tirt_mplus_syntax.R
\name{generate_tirt_mplus_syntax}
\alias{generate_tirt_mplus_syntax}
\title{Generate Mplus Syntax for Thurstonian IRT (TIRT) or Thurstonian Factor Model (TFM)}
\usage{
generate_tirt_mplus_syntax(
  title = "TIRT Model",
  data_file,
  n_blocks,
  block_size,
  key_matrix,
  trait_col,
  key_col,
  cor_matrix = NULL,
  model_type = c("TIRT", "TFM"),
  tfm_free_pair_residuals = TRUE,
  estimator = c("WLSMV", "ULSMV", "ULS"),
  missing_code = "*",
  force_positive_variances = TRUE,
  quick_run = FALSE,
  out_file = "tirt_model.inp"
)
}
\arguments{
\item{title}{Character. Title of the Mplus model.}

\item{data_file}{Character. Name of the input data file.}

\item{n_blocks}{Integer. Number of blocks in the questionnaire.}

\item{block_size}{Integer. Number of items per block.}

\item{key_matrix}{A data.frame with two columns, one indicating item trait,
another indicating item sign.}

\item{trait_col}{Character string. The name of the column in \code{key_matrix} indicating the trait
measured by the item.}

\item{key_col}{Character string. The name of the column in \code{key_matrix} indicating the keying
direction of the item (e.g., positive/negative).}

\item{cor_matrix}{Optional matrix. Starting values for trait correlations.}

\item{model_type}{Character. Choose "TIRT" for the First-Order parameterization
or "TFM" for the Second-Order parameterization.}

\item{tfm_free_pair_residuals}{Logical. In the TFM model, should the residual variance
of the observed pairs be freely estimated? Defaults to TRUE.}

\item{estimator}{Character. The Mplus estimator to use ("WLSMV", "ULSMV", "ULS").}

\item{missing_code}{Numeric or Character. The value representing missing data.}

\item{force_positive_variances}{Logical. Constrain residual variances to be
strictly positive (> 0.001/0) to prevent Heywood cases. Defaults to TRUE.}

\item{quick_run}{Logical. Specify TRUE to speed up estimation by not producing
Chi-Square and standard errors. Defaults to FALSE.}

\item{out_file}{Character. The filename to save the generated Mplus syntax.}
}
\description{
Generate Mplus Syntax for Thurstonian IRT (TIRT) or Thurstonian Factor Model (TFM)
}
