% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_matrices.R
\name{get_simulation_matrices}
\alias{get_simulation_matrices}
\title{Generate Simulated Person and Item Parameter Matrices for the TIRT Model}
\usage{
get_simulation_matrices(cfa_estimates, N, empirical = FALSE)
}
\arguments{
\item{cfa_estimates}{A list object returned by \code{get_CFA_estimates()}.}

\item{N}{Integer. Number of simulated responses (participants) to generate.}

\item{empirical}{Logical. As in \code{MASS::mvrnorm()}: Should mu and sigma
specify the empirical, rather than population, mean and covariance?
Defaults to FALSE.}
}
\value{
A list containing matrices for \code{Lambda} (Loadings), \code{Mu} (Intercepts),
\code{Epsilon} (Residuals), \code{Theta} (Latent Traits), and \code{Utility}.
}
\description{
Takes the factor analysis results extracted by \code{get_CFA_estimates()}
and generates simulated person traits, item parameters, and latent utility values
for a Thurstonian IRT model.
}
\examples{
cfa_model <- paste0("H =~ ", paste0("SS", seq(6, 18, by = 6), collapse = " + "), "\n",
                    "E =~ ", paste0("SS", seq(5, 18, by = 6), collapse = " + "), "\n",
                    "X =~ ", paste0("SS", seq(4, 18, by = 6), collapse = " + "), "\n")
cfa_fit <- lavaan::cfa(cfa_model, data = HEXACO_example_data[1:500,], ordered = TRUE,
                       estimator = "WLSMV", verbose = TRUE, test = "none", se = "none")
cfa_estimates <- get_CFA_estimates(cfa_fit)
simu_estimates <- get_simulation_matrices(cfa_estimates, N = 500)

### Because utilities orders are exactly the order each item appears in the cfa model,
### you may need to re-order the columns back in your own case.

num_order <- order(as.numeric(gsub("[^0-9]", "", colnames(simu_estimates$Utility))))
new_Utility <- simu_estimates$Utility[, num_order]
}
