% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TIRT_long_data.R
\name{get_TIRT_long_data}
\alias{get_TIRT_long_data}
\title{Convert the TIRT pairwise/rank response data into long format, compatible with the thurstonianIRT package}
\usage{
XX
}
\arguments{
\item{block_info}{Information data frame related to keying, dimension, and ID of each item}

\item{response_data}{TIRT pairwise/rank response data.}

\item{response_varname}{Column names of TIRT pairwise/ranked responses. Can be generated from \code{build_TIRT_var_names()}.}

\item{format, direction, family, range}{These parameters works the same as \code{thurstonianIRT::make_TIRT_data()}.}

\item{block_name, item_name, trait_name, sign_name}{These parameters indicate the column names in \code{block_info} that specify
the following information of each item:

\code{block_name}: Which block does this item belong to?
\code{item_name}: What is the name of this item? 
\code{trait_name}: Which trait does this item belong to?
\code{sign_name}: What is the keying of this item?}
}
\value{
A long format dataframe that is compatible with subsequent analyses using the thurstonianIRT package
}
\description{
XX
}
\details{
This function is essentially a wrapper of \code{thurstonianIRT::make_TIRT_data()}
to allow more functionalities to be incorporated in a single function.
}
\examples{
TO BE DONE


}
\seealso{
thurstonianIRT::set_blocks_from_df, thurstonianIRT::make_TIRT_data
}
\author{
Mengtong Li
}
