% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOLE_data.R
\docType{data}
\name{MOLE_data}
\alias{MOLE_data}
\title{Actual Response to FC scales in MOLE format}
\format{
A data frame with 2173 rows (respondents, 4 groups) and 40 columns
#' \describe{
\item{Q1_0_GROUP_T1}{}
\item{Q1_1_GROUP_T1}{}
\item{Q2_0_GROUP_T1}{}
\item{Q2_1_GROUP_T1}{}
\item{Q3_0_GROUP_T1}{}
\item{Q3_1_GROUP_T1}{}
\item{Q4_0_GROUP_T1}{}
\item{Q4_1_GROUP_T1}{}
\item{Q5_0_GROUP_T1}{}
\item{Q5_1_GROUP_T1}{}
\item{Q6_0_GROUP_T1}{}
\item{Q6_1_GROUP_T1}{}
\item{Q7_0_GROUP_T1}{}
\item{Q7_1_GROUP_T1}{}
\item{Q8_0_GROUP_T1}{}
\item{Q8_1_GROUP_T1}{}
\item{Q9_0_GROUP_T1}{}
\item{Q9_1_GROUP_T1}{}
\item{Q10_0_GROUP_T1}{}
\item{Q10_1_GROUP_T1}{}
\item{Q11_0_GROUP_T1}{}
\item{Q11_1_GROUP_T1}{}
\item{Q12_0_GROUP_T1}{}
\item{Q12_1_GROUP_T1}{}
\item{Q13_0_GROUP_T1}{}
\item{Q13_1_GROUP_T1}{}
\item{Q14_0_GROUP_T1}{}
\item{Q14_1_GROUP_T1}{}
\item{Q15_0_GROUP_T1}{}
\item{Q15_1_GROUP_T1}{}
\item{Q16_0_GROUP_T1}{}
\item{Q16_1_GROUP_T1}{}
\item{Q17_0_GROUP_T1}{}
\item{Q17_1_GROUP_T1}{}
\item{Q18_0_GROUP_T1}{}
\item{Q18_1_GROUP_T1}{}
\item{Q19_0_GROUP_T1}{}
\item{Q19_1_GROUP_T1}{}
\item{Q20_0_GROUP_T1}{}
\item{Q20_1_GROUP_T1}{}
\item{Group}{}
}
}
\usage{
MOLE_data
}
\description{
In Li et al's (2025) study, respondents completed one of the four different FC scales.
The current dataset contains their actual responses to the MOLE blocks, as well as their group information.
}
\references{
Li, M., Zhang, B., Li, L., Sun, T., & Brown, A. (2025). Mixed-keying or desirability-matching in the
construction of forced-choice measures? An empirical investigation and practical recommendations.
\emph{Organizational Research Methods, 28}(2), 296-329. \doi{10.1177/10944281241229784}
}
\keyword{datasets}
