% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tirt_stan.R
\name{predict_tirt_stan}
\alias{predict_tirt_stan}
\title{Predict Trait Scores for New Data using a Fitted Stan Model}
\usage{
predict_tirt_stan(fit, stan_data, new_pairwise_data)
}
\arguments{
\item{fit}{A fitted model object from either cmdstanr or rstan.}

\item{stan_data}{The original data list passed to Stan (used to map the test design).}

\item{new_pairwise_data}{A dataframe of pairwise binary responses for the NEW respondents.}
}
\value{
A data frame containing MAP trait scores and Standard Errors.
}
\description{
Extracts posterior means of item parameters from a fitted Stan TIRT
model and uses analytical BFGS optimization to instantly score new respondents.
}
