% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tirt_stan_data.R
\name{prepare_tirt_stan_data}
\alias{prepare_tirt_stan_data}
\title{Prepare pairwise list data for Stan}
\usage{
prepare_tirt_stan_data(
  pairwise_data,
  n_blocks,
  block_size,
  key_matrix,
  trait_col,
  key_col,
  apply_heister = FALSE
)
}
\arguments{
\item{pairwise_data}{A data frame of binary pairwise outcomes (output of
\code{convert_ranks_to_pairwise} or \code{convert_mole_to_pairwise}).}

\item{n_blocks}{Integer. Number of blocks in the questionnaire.}

\item{block_size}{Integer. Number of items per block.}

\item{key_matrix}{A data.frame with two columns, one indicating item trait,
another indicating item sign.}

\item{trait_col}{Character string. The name of the column in \code{key_matrix} indicating the trait
measured by the item.}

\item{key_col}{Character string. The name of the column in \code{key_matrix} indicating the keying
direction of the item (e.g., positive/negative).}

\item{apply_heister}{Logical. Should the optimization techinques proposed by Heister et al., (2025) be applied?
Default is FALSE.}
}
\description{
Prepare pairwise list data for Stan
}
