% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tirt_stan.R
\name{score_tirt_stan}
\alias{score_tirt_stan}
\title{Run Stan TIRT Model and Extract Formatted Scores}
\usage{
score_tirt_stan(
  stan_data,
  chains = 4,
  parallel_chains = 4,
  threads_per_chain = 2,
  iter_warmup = 1000,
  iter_sampling = 1000,
  init = 0
)
}
\arguments{
\item{stan_data}{The data list generated by prepare_tirt_stan_data().}

\item{chains}{Integer. Number of MCMC chains. Defaults to 4.}

\item{parallel_chains}{Integer. How many chains to run simultaneously. Defaults to 4.}

\item{threads_per_chain}{Integer. CPU threads allocated INSIDE each chain. Defaults to 2.}

\item{iter_warmup}{Integer. Warmup iterations. Defaults to 1000.}

\item{iter_sampling}{Integer. Sampling iterations. Defaults to 1000.}

\item{init}{Same as the \code{init} parameter in rstan.}
}
\value{
A list containing \code{scores} (Data frame of traits) and \code{fit} (The cmdstanr fit object).
}
\description{
Run Stan TIRT Model and Extract Formatted Scores
}
