% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_block_info.R
\docType{data}
\name{triplet_block_info}
\alias{triplet_block_info}
\title{Block Information for the Example Triplet Response Data}
\format{
A data frame with 15 rows and 4 columns.
\describe{
  \item{Keying}{Indicates whether the item is positively keyed (1) or negatively keyed (-1).}
  \item{Factor}{Indicates the factor of the item. It can be either one of the six HEXACO traits.}
  \item{Block}{Indicates which block this item belongs to.}
  \item{ID}{Indicates the ID of the item.}
}
}
\source{
https://osf.io/yvpz3/?view_only=08601755f471440b80973194571b60bd
}
\usage{
triplet_block_info
}
\description{
Block design information of the 5-block triplet for \code{triplet_example_data}, 
containing ID, factor, keying, and block membership for the 15 items.
}
\keyword{datasets}
