% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_blueprint.R
\name{construct_blueprint}
\alias{construct_blueprint}
\title{Build an initial blueprint for the FC scale}
\usage{
TO BE DONE
}
\arguments{
\item{N_blocks}{Number of total FC blocks}

\item{block_size}{Desired block size for the FC scale}

\item{traits, signs}{Optional vectors. If given, specifies which traits and signs
each item in the FC scale should have. \code{traits} is a string vector,
while \code{signs} is a numeric vector (1 for positive items and -1 for negative items)}
}
\value{
TO BE DONE
}
\description{
TO BE DONE
}
\details{
TO BE DONE
}
\examples{
TO BE DONE

}
\author{
Mengtong Li
}
