% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_reliability.R
\name{empirical_reliability}
\alias{empirical_reliability}
\title{Calculate the empirical reliability of each latent trait following the formula in Brown & Maydeu-Olivares (2018).}
\usage{
TO BE DONE
}
\arguments{
\item{dataset}{Data frame with trait estimates and standard errors}

\item{score_names}{Vector of characters. Which columns specify trait scores?}

\item{se_names}{Vector of characters. Which columns specify trait standard errors?}
}
\value{
A numeric vector containing empirical reliability estimates, ordered the same as in \code{score_names}.
}
\description{
TO BE DONE
}
\details{
TO BE DONE
}
\examples{
TO BE DONE

}
\author{
Mengtong Li
}
