% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facfun.R
\name{facfun}
\alias{facfun}
\title{Function for Checking If All Items in a Vector Are Unique}
\usage{
facfun(vec)
}
\arguments{
\item{vec}{Input vector.}
}
\value{
\emph{1} if each element in the vector is unique,
and \emph{0} otherwise.
}
\description{
Returns \emph{1} if each element in the vector is unique,
and \emph{0} otherwise.
}
\examples{
  facfun(c("Openness", "Neuroticism", "Agreeableness"))
  facfun(c("Openness", "Openness", "Agreeableness"))
}
\author{
Mengtong Li
}
