% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TIRT_model.R
\name{fit_TIRT_model}
\alias{fit_TIRT_model}
\title{Fit the Thurstonian IRT model with the long format response data and obtain necessary outputs.}
\usage{
XX
}
\arguments{
\item{data_TIRT}{Long format TIRT response data as generated from \code{get_TIRT_long_data} or \code(thurstonianIRT::make_TIRT_data).}

\item{method}{Estimation method for the TIRT model. Can be \code{"lavaan"}, \code{"mplus"} or \code{"stan"}.}

\item{lavaan_estimator}{Which estimator to use when lavaan is chosen as the method of estimating the TIRT model. Defaults to "WLSMV".}

\item{stan_cores, chains, verbose}{Parameters used in \code{thurstonianIRT::fit_TIRT_stan}}
}
\value{
A list containing:
\itemize{
  \code{final_estimates}  Final trait score and standard error estimates

  \code{fit_object}  TIRT model fit object

  \code{responses_TIRT}  The long format TIRT response

  \code{long_estimates}  Final trait score and standard error estimates, in long format

}
}
\description{
TODO
}
\details{
TO BE DONE
}
\examples{
TO BE DONE

}
\seealso{
thurstonianIRT::fit_TIRT_lavaan, thurstonianIRT::fit_TIRT_mplus, thurstonianIRT::fit_TIRT_stan
}
\author{
Mengtong Li
}
