% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_matrices.R
\name{get_simulation_matrices}
\alias{get_simulation_matrices}
\title{Produce simuated matrices based on the Thurstonian IRT model}
\usage{
get_simulation_matrices(loadings, intercepts, residuals, covariances,
                               N, N_items, N_dims, dim_names, empirical)
}
\arguments{
\item{loadings, intercepts, residuals, covariances}{Data frame of factor loadings, intercepts, residuals and latent variable covariances,
preferably obtained from \code{get_CFA_estimates()}, or extracted from \code{lavaan::parameterEstimates()}.}

\item{N}{Number of simulated responses you wish to generate.}

\item{N_items}{Optional. Total number of response items. Default to the number of rows
in \code{loadings}.}

\item{N_dims}{Optional. Total number of response items. Default to the length of \code{dim_names}.}

\item{dim_names}{Name of the latent variables (dimensions); Order should be consistent with 
how they appear in your CFA model as you have specified in \code{get_CFA_estimates()}.}

\item{empirical}{As in \code{MASS::mvrnorm()};  Should mu and sigma specify the empirical, 
rather than population mean and covariance?}
}
\value{
A list containing:
\itemize{
  \code{Lambda} Item loading matrix specifying which items load onto which dimension,

  \code{Mu} Item intercept matrix,

  \code{Epsilon} Item residual matrix,

  \code{Theta} Simulated latent scores for each of the \code{N_dims} dimensions for all \code{N} simualated respondents,
  
  \code{Utility} latent utility value of \code{N_item} Likert items for each of the \code{N} participants.
}
}
\description{
TO BE DONE
}
\details{
Based on the Thurstonian IRT model (Brown & Maydeu-Olivares, 2011), this function
generates the latent utility value of \code{N_item} Likert items for each of the 
\code{N} participants.
}
\examples{
TO BE DONE


}
\seealso{
\code{get_CFA_estimates}
}
\author{
Mengtong Li
}
