% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEXACO_example_data.R
\docType{data}
\name{HEXACO_example_data}
\alias{HEXACO_example_data}
\title{Example HEXACO Response Data}
\format{
A data frame with 100 rows and 60 columns.
\describe{
  \item{SS1, SS2, ... SS60}{Represents the 60 HEXACO items.}
}
}
\source{
https://osf.io/yvpz3/?view_only=08601755f471440b80973194571b60bd
}
\usage{
HEXACO_example_data
}
\description{
Responses to the HEXACO-60 (Ashton & Lee, 2009) scale from 100 actual respondents
}
\references{
Ashton, M. C., & Lee, K. (2009). The HEXACO–60: A short measure of the major dimensions of personality. 
\emph{Journal of Personality Assessment, 91}(4), 340-345. https://doi.org/10.1080/00223890902935878
}
\keyword{datasets}
