% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{Scatter plot for true vs estimated scores, true score vs absolute error, etc.}
\usage{
TO BE DONE
}
\arguments{
\item{x_scores}{Scores to be plotted on the x axis}

\item{y_scores}{Scores to be plotted on the y axis}

\item{type}{Which type of plots is plotted? Can be \code{"simple"} for simple x-y plot, or \code{"abs.diff"}
for plotting absolute difference of (y-x) vs x.}
}
\value{
A scatter plot
}
\description{
TO BE DONE
}
\details{
This is only a very crude plot function extending \code{plot()} for demonstrative purposes.
Users are free to develop their own versions of plotting.
}
\examples{
TO BE DONE

}
\author{
Mengtong Li
}
