% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{Scatter Plot for True vs Estimated Scores, True Score vs Absolute Error, etc.}
\usage{
plot_scores(x_scores, y_scores, type = "simple", ...)
}
\arguments{
\item{x_scores}{Scores to be plotted on the x axis}

\item{y_scores}{Scores to be plotted on the y axis}

\item{type}{Which type of plots is plotted? Can be \code{"simple"} for simple x-y plot, or \code{"abs.diff"}
for plotting absolute difference of (y-x) vs x.}

\item{...}{Other parameters used in \code{plot()}}
}
\value{
A scatter plot
}
\description{
This function is a simple plot for diagnostic purposes examining
the performance of the FC scale based on simulated data.
}
\details{
This is only a very crude plot function extending \code{plot()} for demonstrative purposes.
Users are free to develop their own versions of plotting.
}
\examples{
plot_scores(rnorm(100), rnorm(100))

}
\author{
Mengtong Li
}
