% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_range.R
\name{RMSE_range}
\alias{RMSE_range}
\title{Calculate the overall RMSE of the trait scores, or the RMSE in a certain trait score range}
\usage{
TO BE DONE
}
\arguments{
\item{true_scores}{Actual trait scores}

\item{estimated_scores}{Estimated trait scores from a specified model}

\item{range_breaks}{A numeric vector. Specifies which trait scores ranges will the RMSE be calculated}
}
\value{
TO BE DONE
}
\description{
TO BE DONE
}
\details{
TO BE DONE
}
\examples{
TO BE DONE

}
\author{
Mengtong Li
}
